Renamed the class.
