+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:theme="@style/Theme.CameraXIntegration"
         tools:targetApi="31">
         <activity
-            android:name="com.camera.cameraX.activities.BaseViewPagerActivity"
+            android:name="com.camera.cameraX.activities.CameraActivity"
             android:clearTaskOnLaunch="true"
             android:configChanges="orientation|screenLayout|screenSize|smallestScreenSize"
             android:exported="true"
