+++ app/src/main/java/com/camera/cameraX/fragment/VideoFragment.kt
@@ -29,14 +29,13 @@ import androidx.camera.video.VideoRecordEvent
 import androidx.concurrent.futures.await
 import androidx.core.content.ContextCompat
 import androidx.core.util.Consumer
-import androidx.lifecycle.Observer
 import androidx.lifecycle.lifecycleScope
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import androidx.window.layout.WindowMetricsCalculator
-import com.camera.cameraX.activities.BaseViewPagerActivity.Companion.flashMode
-import com.camera.cameraX.activities.BaseViewPagerActivity.Companion.lensFacing
+import com.camera.cameraX.activities.CameraActivity.Companion.flashMode
+import com.camera.cameraX.activities.CameraActivity.Companion.lensFacing
 import com.camera.cameraX.utils.FILENAME
-import com.camera.cameraX.activities.BaseViewPagerActivity.Companion.telephoneCallReceiver
+import com.camera.cameraX.activities.CameraActivity.Companion.telephoneCallReceiver
 import com.camera.cameraX.utils.FILENAME_FORMAT
 import com.camera.cameraX.utils.TAG
 import com.camera.cameraX.utils.aspectRatio
